src/api/dataflow/requirements.txt
django-filter==2.3.0
django-reversion==2.0.9
enum34==1.1.6
expiringdict==1.2.1
grpcio-tools==1.26.0
grpcio==1.27.1
httplib2==0.7.3
kafka-python==1.4.2
lxml==4.2.5
mock==4.0.3
pika==0.11.0
protobuf==3.14.0
pyahocorasick==1.4.0
pytest-cov==2.12.0
pytest-django==4.3.0
pytest-html==3.1.1
PyYAML==5.4.1
redis-py-cluster==1.3.4
redis==2.10.5
